backend/requirements.txt
fastapi
uvicorn[standard]
pymupdf
python-multipart
groq
python-dotenv
supabase==2.3.5
